added SimInfo and corrected python folder name
